Send object status updates instead of check result
Since each object is owned by a single node, and that node is
responsible of all parameters of the object (except the event queue), we
should blindly copy the status from that node.

Commands executed for an object is already changed to propagate to the
owning node, so nothing should be executed on other nodes.

Thus, we should not send check results, which is interpreted with
process_check_result, since that means we zero out fields like
last_notification.

This patch breaks the scenario of when poller shouldn't notify, but
master should. (Should be fixed in upcoming commits)

Resolves, with exceptions, MON-8720

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>