apps/libexec/confitest.sh: new shortcut to verify config
When we move to systemd unit files the 'service monitor configtest' will
be removed, since that is not part of the init script job to verify.
Systemd is kind of strict on adding more arguments than the defaults.

We also print out command to run to verify from nacoma if your config
is totally broken.

This is part of MON-9855

Signed-off-by: Mattias Ryrlén <mattiasr@op5.com>