Tests: Test block and execution of active checks
This is mostly the foundation of merlin, to be able to distribute checks
over several peers. Thus verify that that works.

It assumes that peer_id is allocated in falling node start time, thus
the latest started node has peer_id=0, the second latest node has
peer_id=1 and so on.

Part of MON-8720

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>