oconf: Extract function for adapting config file to poller
We still replace the contents of a Naemon configuration file, but the
second step of: read config, adapt config, write config is now extracted
to its own function. This makes it easily unit testable.

We also use some more idiomatic Python constructs, such as contexts
("with") that removes the need for manually closing a file handle, and
enumerate() in favor of incrementing a sentinel value manually.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>